feat: layers list displaying connected to store
1. type for Layer instance added
2. layersReducer connected to store
3. separate component for displaying list of
layers added
4. layers are now displayed based on state.layers
