+++ src/components/LayersMenu/LayersMenu.tsx
@@ -1,18 +1,9 @@
-import { Button, Divider, MenuProps, Slider } from 'antd';
-import { Dropdown } from 'antd';
-
-const items: MenuProps['items'] = [
-  {
-    label: 'Переименовать',
-    key: '1',
-  },
-  {
-    label: 'Удалить',
-    key: '2',
-  },
-];
+import { useAppSelector } from '@/redux/hooks';
+import { Divider, Slider } from 'antd';
+import { LayersList } from './LayersList';
 
 export function LayersMenu() {
+  const layers = useAppSelector(state => state.layers);
   return (
     <div className='h-1/2 border-b border-gray-300 '>
       <h1>Layers Menu</h1>
@@ -21,40 +12,7 @@ export function LayersMenu() {
       <h2>Прозрачность</h2>
       <Divider className='m-0' />
       <Slider />
-      <div className='layers'>
-        <div className='flex flex-col justify-between'>
-          <div>
-            <span>Слой 1</span>
-            <button>Скрыть</button>
-            <Dropdown
-              menu={{ items }}
-              placement='bottomLeft'
-              trigger={['click']}>
-              <Button>...</Button>
-            </Dropdown>
-          </div>
-          <div>
-            <span>Слой 2</span>
-            <button>Скрыть</button>
-            <Dropdown
-              menu={{ items }}
-              placement='bottomLeft'
-              trigger={['click']}>
-              <Button>...</Button>
-            </Dropdown>
-          </div>
-          <div>
-            <span>Слой 3</span>
-            <button>Скрыть</button>
-            <Dropdown
-              menu={{ items }}
-              placement='bottomLeft'
-              trigger={['click']}>
-              <Button>...</Button>
-            </Dropdown>
-          </div>
-        </div>
-      </div>
+      <LayersList layers={layers} />
     </div>
   );
 }
